+++ src/constants.rs
@@ -1,0 +1,27 @@
+// Static lorem words to avoid recreation on each call
+pub static LOREM_WORDS: &[&str] = &[
+    "lorem", "ipsum", "dolor", "sit", "amet", "consectetur", "adipiscing", "elit",
+    "sed", "do", "eiusmod", "tempor", "incididunt", "ut", "labore", "et", "dolore",
+    "magna", "aliqua", "enim", "ad", "minim", "veniam", "quis", "nostrud",
+    "exercitation", "ullamco", "laboris", "nisi", "aliquip", "ex", "ea", "commodo",
+    "consequat", "duis", "aute", "irure", "in", "reprehenderit", "voluptate",
+    "velit", "esse", "cillum", "fugiat", "nulla", "pariatur", "excepteur", "sint",
+    "occaecat", "cupidatat", "non", "proident", "sunt", "culpa", "qui", "officia",
+    "deserunt", "mollit", "anim", "id", "est", "laborum"
+];
+
+pub static CAT_FACTS: &[&str] = &[
+    "Cats sleep 12-16 hours a day.",
+    "A group of cats is called a 'clowder'.",
+    "Cats have over 30 muscles controlling their ears.",
+    "A cat's purr vibrates at a frequency that can help heal bones.",
+    "Cats can rotate their ears 180 degrees.",
+];
+
+pub static QUOTES: &[&str] = &[
+    "The only way to do great work is to love what you do. - Steve Jobs",
+    "Innovation distinguishes between a leader and a follower. - Steve Jobs",
+    "Life is what happens to you while you're busy making other plans. - John Lennon",
+    "The future belongs to those who believe in the beauty of their dreams. - Eleanor Roosevelt",
+    "It is during our darkest moments that we must focus to see the light. - Aristotle",
+];
