+++ src/endpoints/mod.rs
@@ -1,0 +1,7 @@
+pub mod basic;
+pub mod crypto;
+pub mod encoding;
+pub mod fun;
+pub mod generators;
+pub mod time;
+pub mod utils;
